rename AtomicGlobalVariables to variables
